# docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "High Street Gym ",
    "description": "JSON REST API for Dynamic Gym website "
  },
  "servers": [
    {
      "url": "http://localhost:8090"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/staff": {
      "post": {
        "summary": "Create a staff",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Adding new staff",
          "content": {
            "application/json": {
              "schema": {
                "email": "string",
                "staff_password": "string",
                "role": "string",
                "phone": "string",
                "first_name": "string",
                "last_name": "string",
                "address": "string"
              },
              "example": {
                "email": "lingCH@server.com",
                "staff_password": "123",
                "role": "admin",
                "phone": "4735658925",
                "first_name": "Jack",
                "last_name": "CHUNG",
                "address": "128 Hassall St"
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all staff",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "summary": "Update an existing staff",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Update an existing staff",
          "content": {
            "application/json": {
              "schema": {
                "authenticationKey": "string",
                "id": "string",
                "email": "string",
                "password": "string",
                "role": "string",
                "firstName": "string",
                "lastName": "string",
                "create_time": "string"
              },
              "example": {
                "authenticationKey": "c2a5afbe-f7a0-4e0a-997d-9cb9241ec52e",
                "id": "6413fc98f808c06fab425866",
                "email": "lingCH333@server.com",
                "password": "lingCH",
                "role": "admin",
                "firstName": "Test",
                "lastName": "staff",
                "create_time": "03/03/2021"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a specific staff by ID",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Deleting a staff",
          "content": {
            "application/json": {
              "schema": {
                "authenticationKey": "string",
                "id": "string"
              },
              "example": {
                "authenticationKey": "c2a5afbe-f7a0-4e0a-997d-9cb9241ec52e",
                "id": "640ec1f2a499175ad989b26c"
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update staff access level by ID ",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Update staff access level ",
          "content": {
            "application/json": {
              "schema": {
                "authenticationKey": "string",
                "role": "string",
                "startDate": "string",
                "endDate": "string"
              },
              "example": {
                "authenticationKey": "c2a5afbe-f7a0-4e0a-997d-9cb9241ec52e",
                "role": "manager",
                "start_date": "01/01/2021",
                "end_date": "01/01/2024"
              }
            }
          }
        }
      }
    },
    "/staff/login": {
      "post": {
        "summary": "staff Login",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "staff Login",
          "content": {
            "application/json": {
              "schema": {
                "email": "string",
                "password": "string"
              },
              "example": {
                "email": "lingCH@server.com",
                "password": "123"
              }
            }
          }
        }
      }
    },
    "/staff/logout": {
      "post": {
        "summary": "staff Logout",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "staff Logout",
          "content": {
            "application/json": {
              "schema": {
                "authenticationKey": "string"
              },
              "example": {
                "authenticationKey": "55d1ebc9-779a-4e2e-8907-184cca7b21f2"
              }
            }
          }
        }
      }
    },
    "/staff/{id}": {
      "get": {
        "summary": "Get a specific staff by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/staff/by-key/{authenticationKey}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authenticationKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
...
}